refactor: Update basic tests to comply with W3C caps format
