+++ test/functional/capabilities.js
@@ -5,13 +5,13 @@ import _ from 'lodash';
 const app = require.resolve('android-apidemos');
 
 const DEFAULT_CAPS = {
-  app,
-  deviceName: 'Android',
+  'appium:app': app,
+  'appium:deviceName': 'Android',
   platformName: 'Android',
 };
 
 const CONTACT_MANAGER_CAPS = _.defaults({
-  app: path.resolve(__dirname, '..', '..', '..', 'test', 'assets', 'ContactManager.apk'),
+  'appium:app': path.resolve(__dirname, '..', '..', '..', 'test', 'assets', 'ContactManager.apk'),
 }, DEFAULT_CAPS);
 
 const CHROME_CAPS = _.defaults({
