+++ test/functional/commands/basic/general-e2e-specs.js
@@ -1,7 +1,7 @@
 import chai from 'chai';
 import chaiAsPromised from 'chai-as-promised';
 import AndroidDriver from '../../../../lib/driver';
-import { DEFAULT_CAPS, CONTACT_MANAGER_CAPS } from '../../desired';
+import { DEFAULT_CAPS, CONTACT_MANAGER_CAPS } from '../../capabilities';
 
 
 chai.should();
