+++ test/functional/commands/find/by-accessibility-id-e2e-specs.js
@@ -1,7 +1,7 @@
 import chai from 'chai';
 import chaiAsPromised from 'chai-as-promised';
 import AndroidDriver from '../../../../lib/driver';
-import DEFAULT_CAPS from '../../desired';
+import DEFAULT_CAPS from '../../capabilities';
 
 
 chai.should();
