+++ test/functional/commands/keyboard/keyboard-e2e-specs.js
@@ -4,7 +4,7 @@ import _ from 'lodash';
 import { retryInterval } from 'asyncbox';
 import AndroidDriver from '../../../../lib/driver';
 import B from 'bluebird';
-import DEFAULT_CAPS from '../../desired';
+import DEFAULT_CAPS from '../../capabilities';
 
 
 chai.should();
